config: recreate the database.yml file to help with ci issues
github actions was having a hard time connecting to the database in order to run the rspec commands.
the funny part of this is that a database is not technically required to run these tests as I'm not saving
any objects to the database at any point. But this will make life easier if I want to extend this application
in the future, potenially with a blog or similar.